Move some MOSS functions to `plagcheck` module
